wip: cleanup ProductsComponent with tonconnect utils
