+++ app/engine/tonconnect/utils.ts
@@ -1,4 +1,99 @@
-import { WebViewBridgeMessageType } from './types';
+import { SEND_TRANSACTION_ERROR_CODES, SessionCrypto } from '@tonconnect/protocol';
+import { Alert } from 'react-native';
+import { Cell } from 'ton';
+import { t } from '../../i18n/t';
+import { getTimeSec } from '../../utils/getTimeSec';
+import { Engine } from '../Engine';
+import { SendTransactionError } from './TonConnect';
+import { SendTransactionRequest, SignRawParams, WebViewBridgeMessageType } from './types';
+
+export const tonConnectTransactionCallback = (
+  ok: boolean,
+  result: Cell | null,
+  request: { from: string } & SendTransactionRequest,
+  sessionCrypto: SessionCrypto,
+  engine: Engine
+) => {
+  if (!ok) {
+      engine.products.tonConnect.send({
+          response: new SendTransactionError(
+              request.id,
+              SEND_TRANSACTION_ERROR_CODES.USER_REJECTS_ERROR,
+              'Wallet declined the request',
+          ),
+          sessionCrypto,
+          clientSessionId: request.from
+      });
+      engine.products.tonConnect.deleteActiveRemoteRequest(request.from);
+      return;
+  }
+
+  engine.products.tonConnect.send({
+      response: { result: result?.toBoc({ idx: false }).toString('base64') ?? '', id: request.id },
+      sessionCrypto,
+      clientSessionId: request.from
+  });
+  engine.products.tonConnect.deleteActiveRemoteRequest(request.from);
+}
+
+export const prepareTonConnectRequest = (request: { from: string } & SendTransactionRequest, engine: Engine) => {
+  const params = JSON.parse(request.params[0]) as SignRawParams;
+
+  const isValidRequest =
+      params && typeof params.valid_until === 'number' &&
+      Array.isArray(params.messages) &&
+      params.messages.every((msg) => !!msg.address && !!msg.amount);
+
+  const session = engine.products.tonConnect.getConnectionByClientSessionId(request.from);
+  if (!session) {
+      engine.products.tonConnect.deleteActiveRemoteRequest(request.from);
+      Alert.alert(t('common.error'), t('products.tonConnect.errors.connection'));
+      return;
+  }
+  const sessionCrypto = new SessionCrypto(session.sessionKeyPair);
+
+  if (!isValidRequest) {
+      engine.products.tonConnect.deleteActiveRemoteRequest(request.from);
+      engine.products.tonConnect.send({
+          response: {
+              error: {
+                  code: SEND_TRANSACTION_ERROR_CODES.UNKNOWN_ERROR,
+                  message: `Bad request`,
+              },
+              id: request.id.toString(),
+          },
+          sessionCrypto,
+          clientSessionId: request.from
+      })
+      return;
+  }
+
+  const { valid_until } = params;
+  if (valid_until < getTimeSec()) {
+      engine.products.tonConnect.deleteActiveRemoteRequest(request.from);
+      engine.products.tonConnect.send({
+          response: {
+              error: {
+                  code: SEND_TRANSACTION_ERROR_CODES.UNKNOWN_ERROR,
+                  message: `Request timed out`,
+              },
+              id: request.id.toString(),
+          },
+          sessionCrypto,
+          clientSessionId: request.from
+      })
+      return;
+  }
+
+  const app = engine.products.tonConnect.findConnectedAppByClientSessionId(request.from);
+
+  return {
+      request,
+      sessionCrypto,
+      messages: params.messages,
+      app
+  }
+}
 
 export const objectToInjection = (obj: Record<string, any>, timeout: number | null) => {
   const funcKeys = Object.keys(obj).filter((key) => typeof obj[key] === 'function');
